LayoutTests/http/tests/paymentrequest/payment-request-change-shipping-option.https-expected.txt

CONSOLE MESSAGE: WebKit currently uses the first shipping option even if other shipping options are marked as selected.

PASS Test for PaymentRequest shippingOption attribute
